requirements.txt
Flask==3.0.0
Flask-JWT-Extended==4.5.3
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
passlib==1.7.4
marshmallow-sqlalchemy==0.29.0
gunicorn==21.2.0
flask-mysql==1.5.2
mysql-connector-python==8.0.5
gevent==23.9.1
eventlet==0.33.3
rarfile==4.1